--- src/main/resources/mappers/order_mapper.xml
@@ -1,43 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!DOCTYPE mapper
-        PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
-        "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
-
-<mapper namespace="com.korit.carecheckkoreait.mapper.OrderMapper">
-    <resultMap id="orderResultMap" type="com.korit.carecheckkoreait.entity.Order">
-        <id property="orderCode" column="order_code" />
-        <result property="orderName" column="order_name" />
-        <result property="orderScore" column="order_score" />
-        <result property="scorePayId" column="score_pay_id" />
-        <result property="orderPay" column="order_pay" />
-        <result property="createAt" column="created_at" />
-        <result property="updateAt" column="updated_at" />
-    </resultMap>
-
-    <insert id="insertOrder">
-        insert into order_tb
-        values(
-            #{orderCode},
-            #{orderName},
-            #{orderScore},
-            default,
-            default,
-            now(),
-            now()
-        )
-    </insert>
-
-    <select id="selectAllByOrderNameContaining" resultMap="orderResultMap">
-        select
-            order_code,
-            order_name,
-            order_score,
-            order_pay
-        from
-            order_tb
-        where
-            order_name like concat("%", #{orderName}, "%")
-    </select>
-
-
-</mapper>
